move /auth/AuthProvider.js -> /controllers/auth.js
