+++ module-9/9.6/client/src/pages/Login.js
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { login, logout } from '../auth/AuthProvider';
+import { login, logout } from '../controllers/auth';
 
 // The login view
 function LoginPage() {
